src/voice/requirements.txt
# Voice service requirements
openai-whisper==20231117
TTS==0.22.0
soundfile==0.12.1
pyaudio==0.2.14
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1
webrtcvad==2.0.10
pydub==0.25.1
pyttsx3==2.90
edge-tts==6.1.9
fastapi==0.104.1
uvicorn==0.24.0